Use current instance method for RPC calls
Previously, this would take a snapshot of all the RPC method coderefs,
and use those when handling any incoming RPC calls.

The downside of this approach is that it leads to surprising behaviour
if you try to override the RPC method with something else later, for
example in a method attribute.

We don't really gain anything by doing this - Perl already caches
method calls. This commit therefore changes the behaviour to a standard
method call keyed by the method name, meaning each RPC call will use
whichever method is in the package symbol table.