Service bus for lighter signalling and notification
This adds a "service bus", accessible through `$api->bus`. It's intended
for notifying other services about events or triggering actions that
don't require the full audit capabilities of a Redis stream.

It uses the sharded Redis pub/sub feature to avoid putting undue load on
larger clusters (the publish/subscribe methods in the transport have been
in place for a while, but used the old publish-to-all-shards PUBLISH/SUBSCRIBE
calls).